修改代码文件名称。

+++ scripts/1_get_pics_from_web_page.py
@@ -5,6 +5,7 @@ import socket
 import json
 from urllib.error import URLError, HTTPError
 import hashlib
+import requests
 
 class MonitorThread(Thread):
 
@@ -72,38 +73,78 @@ def is_valid_jpg(jpg_file):
         f.close()
         return buf ==  b'\xff\xd9'  # 判定jpg是否包含结束字段
 
-def DownloadFile(fullName, url, time=4):
-    opener = urllib.request.build_opener()
-    opener.addheaders = [('User-agent', 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36')]
-    urllib.request.install_opener(opener)
-
+def DownloadFile(path, url, time=5):
+    md5_val = hashlib.md5(url.encode('utf8')).hexdigest()
+    header = {'User-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36'}
     try:
-        urllib.request.urlretrieve(url, fullName)
-        printI('download file %s OK ' % url)
+        s = requests.session()
+        s.keep_alive = False
+        r = s.get(url, headers=header)
+        headers = r.headers
+        #printI(str(headers))
+        if 'Content-Type' in headers:
+            dtype = headers['Content-Type']
+            #printI(dtype)
+            if dtype == 'image/jpeg' or dtype == 'image/png':
+                ext_name = dtype.split('/')[1]
+                pic_full_name = path + '/' + md5_val + '.' + ext_name
+                printI('pic_name : %s, url : %s' % (pic_full_name, url))
+                if os.path.exists(pic_full_name):
+                    printI('file exist %s ' % pic_full_name)
+                    if is_valid_jpg(pic_full_name) is False:
+                        os.remove(pic_full_name)
+                        printI('remove pic')
+                        with open(pic_full_name, 'wb') as f:
+                            f.write(r.content)
+                else:
+                    with open(pic_full_name, 'wb') as f:
+                        f.write(r.content)
+
     except Exception as e:
-        printE('download url : %s ,exception %s' % (url, e))
+        printE('error : %s ' % e)
         if time > 0:
             time -= 1
-            print("DownloadFile %s " % url)
-            DownloadFile(fullName, url, time)
+            printE("DownloadFile faild %s " % url)
+            DownloadFile(path, url, time)
         else:
             printE('file: %s downlaod alway faild' % url)
 
 
-def DownloadPic(path, url_list):
+    # opener = urllib.request.build_opener()
+    # opener.addheaders = [('User-agent', 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/71.0.3578.98 Safari/537.36')]
+    # urllib.request.install_opener(opener)
+    #
+    # try:
+    #     urllib.request.urlretrieve(url, fullName)
+    #     printI('download file %s OK ' % url)
+    # except Exception as e:
+    #     printE('download url : %s ,exception %s' % (url, e))
+    #     if time > 0:
+    #         time -= 1
+    #         print("DownloadFile %s " % url)
+    #         DownloadFile(fullName, url, time)
+    #     else:
+    #         printE('file: %s downlaod alway faild' % url)
+
+
+def DownloadPics(path, url_list):
     for url in url_list:
-        md5_val = hashlib.md5(url.encode('utf8')).hexdigest()
-        print(md5_val, url)
-        pic_full_name = path + '/' + md5_val + '.jpg'
-        #print(pic_full_name)
-        if os.path.exists(pic_full_name):
-            printE('%s exists %s' % (pic_full_name, md5_val))
-            if is_valid_jpg(pic_full_name) is False:
-                printE('%s exists not valid jpg file need download ' % (pic_full_name))
-                os.remove(pic_full_name)
-                DownloadFile(pic_full_name, url)
-        else:
-            DownloadFile(pic_full_name, url)
+        DownloadFile(path, url)
+
+
+    # for url in url_list:
+    #     md5_val = hashlib.md5(url.encode('utf8')).hexdigest()
+    #     print(md5_val, url)
+    #     pic_full_name = path + '/' + md5_val + '.jpg'
+    #     #print(pic_full_name)
+    #     if os.path.exists(pic_full_name):
+    #         printE('%s exists %s' % (pic_full_name, md5_val))
+    #         if is_valid_jpg(pic_full_name) is False:
+    #             printE('%s exists not valid jpg file need download ' % (pic_full_name))
+    #             os.remove(pic_full_name)
+    #             DownloadFile(pic_full_name, url)
+    #     else:
+    #         DownloadFile(pic_full_name, url)
 
 curPath = os.path.abspath('.')
 parentPath = os.path.abspath('..')
@@ -130,7 +171,7 @@ for fileName in fileList:
 
                 after, pic_list = getDownloadPicUrl(jsonString)
                 #print(downloadPath, pic_list)
-                DownloadPic(downloadPath, pic_list)
+                DownloadPics(downloadPath, pic_list)
                 if after is None:
                     printE('url %s after is None' % org_url)
                     break;
